+++ mock.h
@@ -22,13 +22,13 @@ extern uint8_t mock_rd_buffer[];
 
 extern uint32_t handle_counter;
 
-uint8_t mock_dummy_rw(uint8_t* buf, uint32_t count);
+uint32_t mock_dummy_rw(uint8_t* buf, uint32_t count);
 int32_t mock_dummy_handle(uint8_t cmd, uint8_t* data, uint8_t size,
                     uint8_t** rdata, uint8_t* rsize);
-uint8_t mock_reread(uint8_t* buf, uint32_t count);
+uint32_t mock_reread(uint8_t* buf, uint32_t count);
 
-uint8_t mock_read(uint8_t* buf, uint32_t count);
-uint8_t mock_write(uint8_t* buf, uint32_t count);
+uint32_t mock_read(uint8_t* buf, uint32_t count);
+uint32_t mock_write(uint8_t* buf, uint32_t count);
 uint32_t mock_time_ms_func();
 int32_t mock_handle(uint8_t cmd, uint8_t* data, uint8_t size,
                     uint8_t** rdata, uint8_t* rsize);
